+++ CMakeLists.txt
@@ -734,11 +734,7 @@ if(WITH_OPTIM)
         if(WITH_SSE41)
             check_sse41_intrinsics()
             if(HAVE_SSE41_INTRIN)
-                add_definitions(-DX86_SSE41 -DX86_SSE41_ADLER32)
-                set(SSE41_SRCS ${ARCHDIR}/adler32_sse41.c)
-                add_feature_info(SSE4_ADLER32 1 "Support SSE41-accelerated adler32, using \"${SSE41FLAG}\"")
-                list(APPEND ZLIB_ARCH_SRCS ${SSE41_SRCS})
-                set_property(SOURCE ${SSE41_SRCS} PROPERTY COMPILE_FLAGS "${SSE41FLAG} ${NOLTOFLAG}")
+                add_definitions(-DX86_SSE41)
             else()
                 set(WITH_SSE41 OFF)
             endif()
