+++ configure
@@ -1536,10 +1536,8 @@ case "${ARCH}" in
             check_sse41_intrinsics
 
             if test ${HAVE_SSE41_INTRIN} -eq 1; then
-                CFLAGS="${CFLAGS} -DX86_SSE41_ADLER32"
-                SFLAGS="${SFLAGS} -DX86_SSE41_ADLER32"
-                ARCH_STATIC_OBJS="${ARCH_STATIC_OBJS} adler32_sse41.o"
-                ARCH_SHARED_OBJS="${ARCH_SHARED_OBJS} adler32_sse41.lo"
+                CFLAGS="${CFLAGS} -DX86_SSE41"
+                SFLAGS="${SFLAGS} -DX86_SSE41"
             fi
 
             check_sse42_intrinsics
